+++ src/components/mydashboard/NewDashboard.tsx
@@ -3,7 +3,6 @@
 import { getDashboards, postDashboards } from '@/app/api/dashboards';
 import ColorCircle from '@/components/commons/circle/ColorCircle';
 import useDeviceState from '@/hooks/useDeviceState';
-import { colorMapping, colors } from '@/utils/colors';
 import { Dashboard } from '@planit-types';
 import Image from 'next/image';
 import Link from 'next/link';
@@ -14,6 +13,10 @@ import Button from '../commons/button';
 import BarButton from '../commons/button/BarButton';
 import Modal from '../commons/modal';
 
+type ColorMapping = {
+  [key: string]: string;
+};
+
 type ModalState = {
   isOpen: boolean;
   message: string;
@@ -22,6 +25,24 @@ type ModalState = {
 type FormValues = {
   dashboardName: string;
 };
+
+const colorMapping: ColorMapping = {
+  '#5534DA': 'bg-violet-dashboard',
+  '#D6173A': 'bg-red-dashboard',
+  '#7AC555': 'bg-green-dashboard',
+  '#FFA500': 'bg-orange-dashboard',
+  '#76A5EA': 'bg-blue-dashboard',
+  '#E876EA': 'bg-pink-dashboard',
+};
+
+const colors = [
+  '#D6173A',
+  '#E876EA',
+  '#FFA500',
+  '#7AC555',
+  '#5534DA',
+  '#76A5EA',
+];
 
 export default function NewDashboard() {
   const [dashboards, setDashboards] = useState<Dashboard[]>([]);
