--- src/utils/colors.ts
@@ -1,21 +1,0 @@
-type ColorMapping = {
-  [key: string]: string;
-};
-
-export const colorMapping: ColorMapping = {
-  '#5534DA': 'bg-violet-dashboard',
-  '#D6173A': 'bg-red-dashboard',
-  '#7AC555': 'bg-green-dashboard',
-  '#FFA500': 'bg-orange-dashboard',
-  '#76A5EA': 'bg-blue-dashboard',
-  '#E876EA': 'bg-pink-dashboard',
-};
-
-export const colors = [
-  '#D6173A',
-  '#E876EA',
-  '#FFA500',
-  '#7AC555',
-  '#5534DA',
-  '#76A5EA',
-];
